Add files via upload
In this project, we have developed a text – processing system where they compute the TF-IDF scores of different terms within the collection of documents. This process involved tokenizing the text, computing term frequnecy and then calculating inverse document frequency all over the dataset. Therefore we used matrices to compute TF-IDF vectors for each document and applied the cosine similarity to acess the relevance of document to the set of our predefined queries. Then the result were written to a file format for futher analysis. This approach effectively ranks documents according to their relevance to specific queries.